# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"","version":""},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/v1/login":{"get":{"tags":["login"],"summary":"login login","description":"Login endpoint sets up the user for authentication and redirects them to the identity provider","operationId":"login#login","parameters":[{"name":"redirect","in":"query","description":"URL to be redirected to after logging in successfully","required":false,"type":"string"},{"name":"api_client","in":"query","description":"The name of the api client which is requesting a token","required":false,"type":"string"},{"name":"Referer","in":"header","description":"Referer header value","required":false,"type":"string"}],"responses":{"307":{"description":"Temporary Redirect response.","schema":{"$ref":"#/definitions/LoginLoginTemporaryRedirectResponseBody"},"headers":{"Location":{"description":"Location to redirect to for authentication","type":"string"}}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/LoginLoginBadRequestResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/LoginLoginUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/LoginLoginInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/v1/login/callback":{"get":{"tags":["login"],"summary":"callback login","description":"Callback endpoint that receives authorization code from identity provider","operationId":"login#callback","parameters":[{"name":"code","in":"query","description":"Authorization Code","required":false,"type":"string"},{"name":"state","in":"query","description":"State value","required":false,"type":"string"}],"responses":{"307":{"description":"Temporary Redirect response.","schema":{"$ref":"#/definitions/LoginCallbackTemporaryRedirectResponseBody"},"headers":{"Location":{"description":"Location to redirect to for authentication","type":"string"}}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/LoginCallbackBadRequestResponseBody"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/LoginCallbackUnauthorizedResponseBody"}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/LoginCallbackInternalServerErrorResponseBody"}}},"schemes":["http"]}}},"definitions":{"LoginCallbackBadRequestResponseBody":{"title":"LoginCallbackBadRequestResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Voluptas quo."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Eveniet incidunt et ut consequatur."}},"example":{"location":"Modi dolores laborum.","outcome":"Quasi non qui esse non praesentium expedita."}},"LoginCallbackInternalServerErrorResponseBody":{"title":"LoginCallbackInternalServerErrorResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Dignissimos sed eius odio commodi eos consequatur."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Excepturi ut voluptatum."}},"example":{"location":"Aut qui cum temporibus asperiores qui.","outcome":"Quod aut alias optio quidem ducimus."}},"LoginCallbackTemporaryRedirectResponseBody":{"title":"LoginCallbackTemporaryRedirectResponseBody","type":"object","properties":{"outcome":{"type":"string","description":"Login Result Outcome","example":"Officiis est optio voluptates qui recusandae sit."}},"example":{"outcome":"Dolorem eligendi aut consequatur minima rem tempora."}},"LoginCallbackUnauthorizedResponseBody":{"title":"LoginCallbackUnauthorizedResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Aut animi et deserunt est."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Aut et atque."}},"example":{"location":"Ab enim aut et quas quo.","outcome":"Doloremque ut vel."}},"LoginLoginBadRequestResponseBody":{"title":"LoginLoginBadRequestResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Ut molestiae nihil ipsam voluptatem explicabo qui."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Ut ut omnis."}},"example":{"location":"At ad.","outcome":"Quae autem quia nemo iste similique veritatis."}},"LoginLoginInternalServerErrorResponseBody":{"title":"LoginLoginInternalServerErrorResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Ut in sapiente illo explicabo aut."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Magnam non voluptas aut vero pariatur."}},"example":{"location":"Minima nisi.","outcome":"Autem quia veritatis dolorem."}},"LoginLoginTemporaryRedirectResponseBody":{"title":"LoginLoginTemporaryRedirectResponseBody","type":"object","properties":{"outcome":{"type":"string","description":"Login Result Outcome","example":"Aut id placeat voluptate expedita sunt culpa."}},"example":{"outcome":"Voluptatibus cupiditate ea cum ut beatae."}},"LoginLoginUnauthorizedResponseBody":{"title":"LoginLoginUnauthorizedResponseBody","type":"object","properties":{"location":{"type":"string","description":"Location to redirect to for authentication","example":"Nam doloribus dolor commodi consequuntur perferendis ea."},"outcome":{"type":"string","description":"Login Result Outcome","example":"Commodi officiis numquam molestiae."}},"example":{"location":"Earum quia aut.","outcome":"Qui unde et mollitia modi."}}}}